natural_language_processing (2).ipynb
I worked on this project to get insights into how Natural Language Processing is used. This project consists of a naive bayes model that was trained on a bag of words model created through a dataset consisting of positive and negative reviews for a restaurant. After training the model could further predict whether if a future review would be positive or negative.